Conflicts:
	build.cmd

+++ build.cmd
@@ -1,11 +1,8 @@
 @echo off
 
 REM Wipe release directory.
-rmdir /s /q release > nul
-mkdir release
-
-REM Compile the server.
-call coffee -j release\Server.js -c server\Room.coffee server\Game.coffee server\Statistics.coffee server\Player.coffee server\Database.coffee server\Lobby.coffee server\Main.coffee
+IF "%1"=="clean" rmdir /s /q release > nul
+mkdir release 2> nul
 
 REM Install dependencies.
 cd release
@@ -14,5 +11,8 @@ call npm install
 del package.json
 cd ..
 
+REM Compile the server.
+call coffee -j release\Server.js -c server\Room.coffee server\Statistics.coffee server\Game.coffee server\Player.coffee server\Database.coffee server\Lobby.coffee server\Main.coffee
+
 REM Copy client files.
-xcopy /e /i client release\client > nul
+xcopy /y /e /i client release\client > nul
